Added new improvements for progress downloader.

+++ Tollminder.Core/Models/DownloadBytesProgress.cs
@@ -4,16 +4,16 @@ namespace Tollminder.Core.Models
 {
 	public class DownloadBytesProgress
 	{
-		public DownloadBytesProgress(string fileName, int bytesReceived, int totalBytes)
+		public DownloadBytesProgress(string fileName, long bytesReceived, long totalBytes)
 		{
 			Filename = fileName;
 			BytesReceived = bytesReceived;
 			TotalBytes = totalBytes;
 		}
 
-		public int TotalBytes { get; private set; }
+		public long TotalBytes { get; private set; }
 
-		public int BytesReceived { get; private set; }
+		public long BytesReceived { get; private set; }
 
 		public float PercentComplete { get { return ((float)BytesReceived / TotalBytes) * 100; } }
